Add README and future modules
All of the files for this pre-work were being displayed in my staging area, which made it difficult to find the desired
exercise I wanted to stage, commit, and push.  So this commit is a bulk staging of all the empty, future exercises that I
have not completed yet.  They will just display in my repository as blank files for now, but will have content in future commits.

Stay tuned!