Replace external link with internal link
I changed the external link to an internal link to my footer section by adding an id value to the footer element.
The href value of the anchor element was changed to "#footer" which tells the browser to navigate to the footer
section of the page when the link is clicked.

"target=_blank" was also removed from the anchor element because I don't want to open a new window when the
link is clicked.