Add inline style to h2 element
I wanted to change the heading color to red, so I applied an inline style to the <h2> tag.

Since I'm using an inline style as opposed to an external style sheet, I changed the file extention to .html so the text editor doesn't throw errors.

I also changed the comments to html format for consistency with the file extension.