Assign CSS class to an element
Set up a class declaration to apply to multiple elements to assign the same style.  The benefit is less code to write.

I also realized that I can change the language mode at the bottom of VS Code instead of changing the file
extension to handle the errors being read by the IDE.  From this point, I will leave the file extension as .css and just change the language mode
to HTML.